fix: refactoring the code for thruster allocation
