+++ controllers/__init__.py
@@ -1,0 +1,2 @@
+from dsorlib.controllers.inner_loops.pid import PID
+from dsorlib.controllers.inner_loops.auv_inner_loops import yaw_rate_PI, yaw_PD, surge_PI
