+++ vehicles/__init__.py
@@ -1,1 +1,2 @@
-from . import thrusters
+from .state import state
+from . import thrusters, auv_dynamics, ocean_currents
