+++ vehicles/auv_dynamics/__init__.py
@@ -1,0 +1,8 @@
+# Import the abstract class that is the blueprint for implementing AUV dynamics
+from .abstract_auv_dynamics import AbstractAUVDynamics
+
+# Import the class that can implement a neutral buoyancy vehicle in an efficient manner
+from .neutral_buoyancy_auv_dynamics import NeutralBuoyancyAUVDynamics
+
+# Import the class where the vehicle is approximated by a sphere (for buoyancy purposes)
+from .sphere_auv_dynamics import SphereAUVDynamics
