+++ vehicles/auv_dynamics/abstract_auv_dynamics.py
