+++ vehicles/auv_dynamics/sphere_auv_dynamics.py
@@ -21,7 +21,7 @@
 #  SOFTWARE.
 from numpy import array, power, pi, sin, cos
 
-from dsorlib.vehicles.auv_rb_dynamics.abstract_auv_dynamics import AbstractAUVDynamics
+from dsorlib.vehicles.auv_dynamics.abstract_auv_dynamics import AbstractAUVDynamics
 from dsorlib.vehicles.state.state import State
