+++ vehicles/thrusters/__init__.py
@@ -1,0 +1,7 @@
+# Import the thrusters
+from .thruster import Thruster
+from .simple_pole_thruster import SimplePoleThruster
+
+# Import the thruster allocators
+from .thruster_allocater import ThrusterAllocator
+from .static_thruster_allocator import StaticThrusterAllocator
