+++ vehicles/thrusters/deprecated/abstract_thruster_model.py
