+++ vehicles/thrusters/thruster_allocater.py
@@ -1,34 +1,49 @@
+#  MIT License
+#
+#  Copyright (c) 2021 Marcelo Jacinto
+#
+#  Permission is hereby granted, free of charge, to any person obtaining a copy
+#  of this software and associated documentation files (the "Software"), to deal
+#  in the Software without restriction, including without limitation the rights
+#  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+#  copies of the Software, and to permit persons to whom the Software is
+#  furnished to do so, subject to the following conditions:
+#
+#  The above copyright notice and this permission notice shall be included in all
+#  copies or substantial portions of the Software.
+#
+#  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+#  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+#  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+#  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+#  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+#  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+#  SOFTWARE.
 from abc import ABC, abstractmethod
-from numpy import ndarray, array, zeros, dot
+from numpy import ndarray, array, zeros, dot, clip
 
 from dsorlib.vehicles.thrusters.thruster import Thruster
 
-# Type to ensure that the list passed to the ThrusterAllocater class is actual thrusters
+# Type to ensure that the list passed to the ThrusterAllocator class is actual thrusters
 ThrustersList = list[Thruster]
 
 
 class ThrusterAllocator(ABC):
 
-    def __init__(self,
-                 number_of_thrusters: int,
-                 thrusters: ThrustersList):
+    def __init__(self, thrusters: ThrustersList):
         """
         The constructor for the abstract class ThrusterAllocator. The goal of this class is to provide
         an interface for thruster allocation
-        :param number_of_thrusters: The number of thrusters that the vehicle has
         :param thrusters: A list of thrusters objects
         """
         if type(self) == ThrusterAllocator:
             raise Exception("<ThrusterAllocator> cannot be instantiated. It is an abstract class")
 
         # Save and check the inputs received by the constructor
-        self.number_of_thrusters: int = int(number_of_thrusters)
+        self.number_of_thrusters: int = len(thrusters)
 
         # Check if the list of thrusters contains actual thruster objects
         self.thrusters = thrusters
-
-        if len(self.thrusters) != self.number_of_thrusters:
-            raise Exception("The number of thruster is not the same as the length of the thrusters list.")
 
     def _thrust_to_force(self, B: ndarray, thrusts: ndarray):
         """
@@ -65,6 +80,10 @@ class ThrusterAllocator(ABC):
 
             # Convert the force in [N] to the input unit of the system (for example RPM)
             input_value = thruster.force_to_input_unit(thrusts[i])
+
+            # Saturate the input of the system between the minimum and maximum values
+            min_val, max_val = thruster.input_bounds
+            input_value = clip(input_value, min_val, max_val)
 
             # Apply the dynamics of that thruster
             actual_applied_value = thruster.dynamic_model(input_value)
